requirements.txt
google-api-python-client
jieba
torch
transformers
datasets
matplotlib
wordcloud
pillow
numpy
tqdm
